requirements.txt
openai==1.12.0
pinecone-client==3.0.3
pandas==2.1.4
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
numpy==1.24.3
tqdm==4.66.1
